pyproject.toml
[project]
name = "trailer-edit"
version = "0.1.0"
description = "Add your description here"
readme = "readme.md"
requires-python = ">=3.10"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "scenedetect[opencv]>=0.6.6",
]
